Update patch set 1
Patch Set 1: Code-Review+1

When this change is coupled with http://review.cyanogenmod.org/#/c/53131/, the network mode can be toggled without issue between 4G and the other values. Without the services/Telephony change, if the network mode is ever set to any value other than 7, it can never be set to 7 again.

Patch-set: 1