Update patch set 1
Patch Set 1: Code-Review-1

This can't go in yet as I'm pretty sure it would break automated builds that already have their props extracted on the filesystem. I'll take a closer look this week to see if I'm wrong about that, or if I need to make a tweak to either this or the existing props to make it work.

Patch-set: 1